requirements.txt

fastapi==0.109.1
uvicorn==0.27.1
google-generativeai==0.3.2
python-dotenv==1.0.1
pydantic==2.6.1
librosa==0.10.1
numpy==1.26.4
